pyproject.toml
[tool.poetry]
name = "nas rental home price prediction"
version = "0.1.0"
description = "nas rental home price prediction machine learning model"
authors = ["nesruler <72273305+boss2256@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.2"
pycaret = "^3.3.0"
pandas = "^2.2.1"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
